requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
crispy-bootstrap4==2023.1
dj-database-url==2.1.0
Django==4.2.10
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
django-heroku==0.3.1
django-pwa==1.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
python-decouple==3.8
pytz==2024.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
